BehaviourComposer passes the parameters for Status message to rpc module
