+++ src/behaviour.rs
@@ -1,6 +1,7 @@
 use crate::discovery::behaviour::DiscoveryEvent;
 use crate::peer_manager::PeerManagerEvent;
 use crate::rpc::behaviour::RpcEvent;
+use crate::types::{default_finalized_root, Epoch, Root, Slot};
 use libp2p::swarm::handler::DummyConnectionHandler;
 use libp2p::swarm::{NetworkBehaviourAction, NetworkBehaviourEventProcess, PollParameters};
 use libp2p::NetworkBehaviour;
@@ -92,9 +93,17 @@ impl NetworkBehaviourEventProcess<PeerManagerEvent> for BehaviourComposer {
                 // The dialing client MUST send a Status request upon connection.
                 // https://github.com/ethereum/consensus-specs/blob/dev/specs/phase0/p2p-interface.md#status
 
+                // TODO: Fill the fields with the real values
                 // ref: Building a `StatusMessage`
                 // https://github.com/sigp/lighthouse/blob/4bf1af4e8520f235de8fe5f94afedf953df5e6a4/beacon_node/network/src/router/processor.rs#L374
-                self.rpc.send_status(peer_id);
+                self.rpc.send_status(
+                    peer_id,
+                    [0; 4],
+                    default_finalized_root(),
+                    Epoch::new(0),
+                    Root::from_low_u64_le(0),
+                    Slot::new(0),
+                );
             }
         }
     }
