+++ src/rpc/message.rs
@@ -1,27 +1,4 @@
-use primitive_types::H256;
-
-// `compute_fork_digest(current_fork_version, genesis_validators_root)`
-type ForkDigest = [u8; 4];
-
-pub(crate) type Root = H256;
-
-#[derive(Debug)]
-pub(crate) struct Epoch(u64);
-
-impl Epoch {
-    pub(crate) fn new(n: u64) -> Epoch {
-        Epoch(n)
-    }
-}
-
-#[derive(Debug)]
-pub(crate) struct Slot(u64);
-
-impl Slot {
-    pub(crate) fn new(n: u64) -> Slot {
-        Slot(n)
-    }
-}
+use crate::types::{Epoch, ForkDigest, Root, Slot};
 
 // /////////////////////////////////////////////////////////////////////////////////////////////////
 // `Status` request/response handshake message.
@@ -43,9 +20,4 @@ pub(crate) struct Status {
     pub(crate) head_root: Root,
     // The slot of the block corresponding to the `head_root`.
     pub(crate) head_slot: Slot,
-}
-
-// `finalized_root` defaults to Root(b'\x00' * 32) for the genesis finalized checkpoint
-pub(crate) fn default_finalized_root() -> Root {
-    Root::from_low_u64_le(0)
 }
